+++ LabelStoreMax/lib/app/providers/firebase_provider.dart
@@ -1,5 +1,6 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
+import 'package:flutter_app/bootstrap/app_helper.dart';
 import 'package:flutter_app/firebase_options.dart';
 import 'package:nylo_framework/nylo_framework.dart';
 import 'package:woosignal/woosignal.dart';
@@ -14,7 +15,12 @@ class FirebaseProvider implements NyProvider {
 
   @override
   afterBoot(Nylo nylo) async {
-    if (getEnv('FCM_ENABLED', defaultValue: false) != true) return;
+    bool? firebaseFcmIsEnabled = AppHelper.instance.appConfig?.firebaseFcmIsEnabled;
+    if (firebaseFcmIsEnabled == null) {
+      firebaseFcmIsEnabled = getEnv('FCM_ENABLED', defaultValue: false);
+    }
+
+    if (firebaseFcmIsEnabled != true) return;
 
     await Firebase.initializeApp(
       options: DefaultFirebaseOptions.currentPlatform,