+++ LabelStoreMax/lib/firebase_options.dart
@@ -2,6 +2,7 @@
 import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
 import 'package:flutter/foundation.dart'
     show defaultTargetPlatform, kIsWeb, TargetPlatform;
+import 'package:flutter_app/bootstrap/app_helper.dart';
 
 /// Default [FirebaseOptions] for use with your Firebase apps.
 class DefaultFirebaseOptions {
@@ -14,9 +15,33 @@ class DefaultFirebaseOptions {
     }
     switch (defaultTargetPlatform) {
       case TargetPlatform.android:
-        return android;
+        if (AppHelper.instance.appConfig?.firebaseOptionsAndroid == null) {
+          throw UnsupportedError(
+            'Add a valid Firebase json config on https://woosignal.com for your WooCommerce store',
+          );
+        }
+        return FirebaseOptions(
+          apiKey: AppHelper.instance.appConfig!.firebaseOptionsAndroid!['apiKey'],
+          appId: AppHelper.instance.appConfig!.firebaseOptionsAndroid!['appId'],
+          messagingSenderId: AppHelper.instance.appConfig!.firebaseOptionsAndroid!['messagingSenderId'],
+          projectId: AppHelper.instance.appConfig!.firebaseOptionsAndroid!['projectId'],
+          storageBucket: AppHelper.instance.appConfig!.firebaseOptionsAndroid!['storageBucket'],
+        );
       case TargetPlatform.iOS:
-        return ios;
+        if (AppHelper.instance.appConfig?.firebaseOptionsIos == null) {
+          throw UnsupportedError(
+            'Add a valid Firebase plist config on https://woosignal.com for your WooCommerce store',
+          );
+        }
+        return FirebaseOptions(
+          apiKey: AppHelper.instance.appConfig!.firebaseOptionsIos!['apiKey'],
+          appId: AppHelper.instance.appConfig!.firebaseOptionsIos!['appId'],
+          messagingSenderId: AppHelper.instance.appConfig!.firebaseOptionsIos!['messagingSenderId'],
+          projectId: AppHelper.instance.appConfig!.firebaseOptionsIos!['projectId'],
+          storageBucket: AppHelper.instance.appConfig!.firebaseOptionsIos!['storageBucket'],
+          iosClientId: AppHelper.instance.appConfig!.firebaseOptionsIos!['iosClientId'],
+          iosBundleId: AppHelper.instance.appConfig!.firebaseOptionsIos!['iosBundleId'],
+        );
       case TargetPlatform.macOS:
         throw UnsupportedError(
           'DefaultFirebaseOptions have not been configured for macos - '
@@ -38,22 +63,4 @@ class DefaultFirebaseOptions {
         );
     }
   }
-
-  static const FirebaseOptions android = FirebaseOptions(
-    apiKey: '',
-    appId: '',
-    messagingSenderId: '',
-    projectId: '',
-    storageBucket: '',
-  );
-
-  static const FirebaseOptions ios = FirebaseOptions(
-    apiKey: '',
-    appId: '',
-    messagingSenderId: '',
-    projectId: '',
-    storageBucket: '',
-    iosClientId: '',
-    iosBundleId: '',
-  );
 }